PROJ-40 refactor: change name for constants folder
