+++ shared/api/config.ts
@@ -1,2 +1,1 @@
-export const BASE_URL_API =
-  process.env.NEXT_PUBLIC_BASE_URL_API || 'https://auth.tehnom.org/api/v1/'
+export const BASE_URL_API = process.env.NEXT_PUBLIC_BASE_URL_API
