+++ shared/constants/auth.ts
